Change definitions to match declarations
Both CalcGravitationalConstant and CalcHubbleUnit are declared as
returning a Double_t, but the were defined as returning double. This
works fine if Double_t and double are the same type, but fails
otherwise.

There are two ways to go about this: either change the declaration to
return double, or the definition to return Double_t. I opted for the
latter, as the operands are already of Double_t type.

This issue was originally reported in #39.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>