Simplify safe_hdf5 function usage
The previous incarnation of this function required callers to specify
the function return type, but safe_hdf5 already has all the information
required to automatically calculate this.

This new version of safe_hdf5 automatically calculates the return type,
and also prints the hdf5 function name in the error message. To do so
without much friction safe_hdf5 is actually a macro that feeds
_safe_hdf5, the actual function.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>