Unify write_attribute method implementations
There were two very similar implementations of this functionality, one
for plain scalar values and another one for string, but at the end they
were very much the same.

This commit puts the common bits of both implementations in a private
method, then adjusts the two existing public methods to invoke the
private one with the correct arguments.

In addition to this unification, the version of write_attribute taking a
string value was failing to close the new data type it created, leading
to object identifier leaks in HDF5. This has been now fixed.

Finally, scalar attributes can (and are, in VR) only be written in
serial. A new assertion to this effect has been added to the code.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>